move google map to dir
