+++ front-end/src/main.jsx
@@ -2,6 +2,6 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
-import './embla.css'
+// import './embla.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(<App />)
