+++ front-end/src/pages/MainMapGoogle/LocationMarker.jsx
